precarga de oferta y subasta

+++ Obligatorio1/Dominio/Entidades/Oferta.cs
@@ -6,12 +6,12 @@ namespace Dominio.Entidades
     {
         public int Id { get; set; }  
         public int IdUser { get; set; }
-        public decimal Monto { get; set; }
+        public double Monto { get; set; }
         public DateTime FchOfer { get; set; }
         public string Pnombre { get; set; }
         private static int _ultimoId;
 
-        public Oferta(int id, int idUser, decimal monto, DateTime fchOfer, string pnombre)
+        public Oferta(int id, int idUser, double monto, DateTime fchOfer, string pnombre)
         {
             Id = _ultimoId++;
             IdUser = idUser;
